requirements.txt
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
facebook-sdk==3.0.0
idna==2.7
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==3.0.0
mccabe==0.6.1
numpy==1.15.2
pandas==0.23.4
pylint==2.1.1
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
wrapt==1.10.11
